requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
# Optional dependencies for AI features
boto3>=1.28.0
botocore>=1.31.0
# strands  # Uncomment if you want AI recommendations